Add JNI header files generated from jaggl and jagmisc
A long-term goal is to write clean room implementations of these
libraries, for several reasons:

 * The original libraries are missing, so we currently have to do a fair
   amount of awkward bytecode patching in the loader to use JOGL
   instead.
 * The original libraries were not compiled for 64-bit Linux. We might
   also want to compile for new architectures (e.g. ARM or RISC-V) or
   additional operating systems (e.g. the BSDs).
 * I'm worried that over time the libraries will bit rot and might
   eventually stop working with modern JREs - it's already been about 10
   years since they were originally compiled!

I'm not too concerned about putting this in the main open-source
repository: I think it's generally believed that interfaces are not
copyrightable, and we're only copying the interface for
interoperability purposes. The native code will be original.

Furthermore, most of the jaggl interface is just a copy of the OpenGL
interface, so it's unlikely that Jagex could claim they hold the
copyright on it anyway.