Skip merging logic in IntInterpreter if value1 and value2 are equal
This fixes a problem where dummy arguments that already had more than 8
possible values in their set (because the dummy argument is used in a
greater or less than condition) were not removed, as we'd replace the
IntValue with an unknown value when two branches (where the dummy
argument is not modified) are merged back together.

It should also improve performance, as we'll be doing less memory
allocation.