Remove mutually-recursive dummy method calls
This is a little bit grim and probably not completely safe in all cases,
but it works well enough on the client.

Ideally I think I'd do it with a dominator tree calculated from a call
graph aware of integer constants and conditional calls, but that's quite
complicated (especially given how the existing code in the
DummyArgTransformer works).