Add openrs2 prefix to repo{Username,Password} properties
These might be added to the user's ~/gradle.properties file, at which
point having a per-project prefix is useful.

Ideally I'd have preferred to use openrs2.repo.{username,password}.
However, Gradle properties with dots in them can't be configured with
the ORG_GRADLE_PROJECT_* environment variables, which we use in the
Jenkinsfile.