Account for inheritance in StaticScramblingTransformer
GETSTATIC, PUTSTATIC and INVOKESTATIC (strangely) use the same field and
method resolution algorithms as their non-static equivalents.

While this doesn't cause problems for 550, as the *STATIC instructions
seem to always refer to the class containing the static definition,
Major reports that OSRS clients sometimes change the owner to a
subclass. This broke the old transformer.

This commit fixes it by applying the old to new owner mapping to an
entire disjoint set at a time.

Signed-off-by: Graham <gpe@openrs2.dev>