Switch to JDK11
I'm very keen on being able to use the jdk.jartool module (which is only
available in JDK11 onwards) as it allows us to avoid shelling out to
jarsigner entirely.

11 is the current LTS release and is already widespread in Linux
distributions, so I think it's reasonable to require it.

This commit removes the jsobject module. We might need to re-add it in
the future (if jdk.jsobject is removed from the JDK). However, it was
only necessary in 8 because modern versions of 8 tended to be
distributed without plugin.jar. JDK11 is distributed with the
jdk.jsobject module.

Signed-off-by: Graham <gpe@openrs2.dev>