Set artifactId of all submodule to openrs2
Now we don't use Maven, the dev.openrs2:openrs2 Maven coordinates are
available for use. I think using them for the end-user application
packaged as a .zip file is reasonable.

I don't intend to move the all submodule to the root build.gradle.kts
file, as this would cause us to lose some of the benefits of supporting
Gradle configuration on demand.

Signed-off-by: Graham <gpe@openrs2.dev>