Only use StackFrameClassWriter where required
Java 11's Pack200 implementation complains if stack map frames are present, and
passes the classes through unmodified. I'm not sure why.

Stack map frames are only required in Java 7, so for 6 and below a reasonable
workaround is to not emit them.

A nice side effect of this change is that the jar files created by the bundler
are be smaller. The pack200 files are too, and would be even if the classes
were not passed through unmodified.