Rename private.key to game.key
This name is slightly more descriptive and consistent with loader.p12,
which also contains a private key.

I've also removed the public.key file. It isn't read by OpenRS2 and the
two files could get out of sync with each other, which would be
confusing.

It's easy to extract the public key with OpenSSL if necessary:

  openssl pkey -in game.key -pubout -out public.key

Signed-off-by: Graham <gpe@openrs2.dev>