Add getFieldInitializer to ExtendedRemapper
My rough plan for combining static scrambling and remapping is to split
Library::remap into three passes:

* Remove static methods, fields and initializers from the library,
  storing them in a temporary location.
* Pass all classes through ClassNodeRemapper, as we do now.
* Add static methods, fields and initializers to their new classes,
  remapping as we do so.

This ensures a ClassNode is never in a state where it has a mixture of
remapped and non-remapped fields, methods or instructions. This is
important to ensure no conflicts can occur when we use the refactored
names from the NameMap, rather than the auto-generated names.

It means TypedRemapper needs the ability to provide the instructions
that make up a field's initializer, such that Library::remap can move
these instructions to a different InsnList. The new getFieldInitializer
method and FieldInitializer type support this.

Signed-off-by: Graham <gpe@openrs2.dev>