Rename Js5Library{Reader,Writer} to PackClassLibrary{Reader,Writer}
JS5 seems to refer to many things, so I'm intending to name Jagex's
custom class file format "packclass" instead of "js5". My reasoning is
that the tool for unpacking it is called unpackclass. If you use the
same logic behind naming the Pack200 packer unpack200, then "packclass"
is a reasonable name for Jagex's class file format.

Signed-off-by: Graham <gpe@openrs2.dev>