Always return a mutable Int2ObjectSortedMap from Group.unpack()
Originally I used a singleton map if possible, to reduce overhead in the
common case of a group containing a single file. However, the
work-in-progress Cache class needs to mutate the map.

Eventually my plan is to replace the Int2ObjectSortedMaps in Js5Index
and Group with arrays, solving both the overhead and mutation problems.

Signed-off-by: Graham <gpe@openrs2.org>