Set background colour of the indexes, groups and keys cells
Green indicates we've collected a full set. For indexes, red indicates
some are missing (as this is a critical problem - the client won't start
at all if an index is missing). Yellow indicates groups or keys are
missing, as this is less critical (the client will likely work in most
cases).

Signed-off-by: Graham <gpe@openrs2.org>