Ignore trailing blocks with a non-zero next block pointer
While the client doesn't appear to be capable of producing a trailing
block with a non-zero next block pointer (though I may have misread the
code), there are caches out there in the wild with trailing non-zero
next block pointers.

When it is reading a group, the client ignores these.

Therefore, for compatibility with the client and existing caches, this
commit removes the StoreCorruptException thrown in this case.

Signed-off-by: Graham <gpe@openrs2.org>