Fix LEFT JOINs in queries for listing missing groups
The previous code included too many rows as the use of LEFT JOINs meant
candidate group orws with an incorrect container_id were still included
in the results.

Using an IN clause with a subquery allows us to remove those rows,
though it's a bit hacky.

(Really I want to be able to use the JOIN on the right side of a LEFT
JOIN to restrict the rows that appear in the results of the LEFT JOIN,
but that doesn't seem to be possible.)

This is similar to the issue fixed by
a920570f042ff27e476429d1caaea69dce2120fa.

Signed-off-by: Graham <gpe@openrs2.org>