Add resolve_{group,index} functions
These functions reduce the amount of group resolution logic
significantly, concentrating it in a single place. In addition to the
usual code de-duplication benefits, many of the queries are now much
simpler as the complexity is hidden behind the function calls.

This change also allows us to make the group resolution logic more
complicated. The first change is that the functions are guaranteed to
only return a single row, which was not true with the old JOIN-based
approach. The row that is chosen is chosen deterministically.

The resolution logic will probably be improved in the future, so we can
make a better decision where there are multiple possible groups, due to
collisions.

Signed-off-by: Graham <gpe@openrs2.org>