Add lax gzip implementation
This is closer to the client's implementation, which ignores the gzip
header/trailer entirely and just decompresses the inner DEFLATE stream.

While Jagex always produce valid gzip files, some cache editors in the
private server scene do not set the checksum/length correctly in the
trailer. I'm planning to add an option to use the lax gzip
implementation to support reading these caches.

Signed-off-by: Graham <gpe@openrs2.org>