Move length encoding/decoding from Rs2{Decoder,Encoder} to PacketCodec
There are now three additional abstract PacketCodec classes:
FixedPacketCodec, VariableBytePacketCodec and VariableShortPacketCodec.

The PacketLength class has been removed, as it is no longer required.

The main reason for this change is that the create suggested names
packet is a bit of an oddball: its length field measures the size of the
packet in longs, not bytes. The codec for this packet will be able to
inherit from PacketCodec directly to implement the custom length logic.

Signed-off-by: Graham <gpe@openrs2.org>