Add Arena-based ByteBuf allocator
This isn't quite true arena allocation, as we're forwarding all requests
to an underlying ByteBufAllocator, which is likely used for other
allocations too. Under the hood it still frees all the allocations
individually, however, to the programmer it allows them to be freed in
one go. This makes writing code that uses a large number of buffers in
the same scope cleaner - particularly in encoders for file formats that
use striped encoding, like the packclass and song formats.

Signed-off-by: Graham <gpe@openrs2.org>