Define deobfuscation modules in profile
The deobfuscation commands now load the modules to operate on from the
profile, allowing other client builds to be deobfuscated without code
edits. The previously-hardcoded modules are now specified in the default
profile, and multiple modules of the same type are not supported.

The BytecodeDeobfuscator still contains a small amount of build-specific
logic, but that now only executes if the appropriate modules are
declared in the profile.

Signed-off-by: Soph <73026338+sophxm@users.noreply.github.com>