Defer node removal in IdentityTransformer
Javaparser does not appear to support removing nodes from the AST
mid-traversal, only replacing them. With the previous behaviour, if a
statement in the walk is redundant (e.g. `anInt123 += 0;`), it will
be removed and Javaparser will skip over the next node (as its internal
index is now off by one) or IndexOutOfBounds (if the removed statement
was the final one in the walk).

Signed-off-by: Soph <73026338+sophxm@users.noreply.github.com>